+++ amfora.go
@@ -10,7 +10,7 @@ import (
 )
 
 var (
-	version = "1.5.0"
+	version = "v1.6.0"
 	commit  = "unknown"
 	builtBy = "unknown"
 )