+++ display/download.go
@@ -117,10 +117,12 @@ func dlChoice(text, u string, resp *gemini.Response) {
 			portalURL = parsed.String() + "%3F" + query
 		}
 		portalURL = strings.TrimPrefix(portalURL, "gemini://") + "?raw=1"
-		handleHTTP("https://portal.mozz.us/gemini/"+portalURL, false)
-		tabPages.SwitchToPage(strconv.Itoa(curTab))
-		App.SetFocus(tabs[curTab].view)
-		App.Draw()
+		ok := handleHTTP("https://portal.mozz.us/gemini/"+portalURL, false)
+		if ok {
+			tabPages.SwitchToPage(strconv.Itoa(curTab))
+			App.SetFocus(tabs[curTab].view)
+			App.Draw()
+		}
 		return
 	}
 	tabPages.SwitchToPage(strconv.Itoa(curTab))